refactor: change the directory structure of the network test program
