+++ tests/contest/contest/src/main.rs
@@ -110,9 +110,9 @@ fn main() -> Result<()> {
     let cgroup_v1_cpu = cgroups::cpu::v1::get_test_group();
     let cgroup_v2_cpu = cgroups::cpu::v2::get_test_group();
     let cgroup_v1_memory = cgroups::memory::get_test_group();
-    let cgroup_v1_network = cgroups::network::get_test_group();
+    let cgroup_v1_network = cgroups::network::network::get_test_group();
     let cgroup_v1_blkio = cgroups::blkio::get_test_group();
-    let cgroup_v1_relative_network = cgroups::relative_network::get_test_group();
+    let cgroup_v1_relative_network = cgroups::network::relative_network::get_test_group();
     let seccomp = get_seccomp_test();
     let seccomp_notify = get_seccomp_notify_test();
     let ro_paths = get_ro_paths_test();
