+++ tests/contest/contest/src/tests/cgroups/mod.rs
@@ -9,7 +9,6 @@ pub mod cpu;
 pub mod memory;
 pub mod network;
 pub mod pids;
-pub mod relative_network;
 
 pub fn cleanup_v1() -> Result<()> {
     for subsystem in list_subsystem_mount_points()? {
